+++ PresentationLayer/Models/IModel.vb
@@ -89,7 +89,7 @@ Public Interface IModel
 
     Function GetSortedRecordPosition(idNo As Int32, tableName As String, sortOrder As String) As Integer
 
-    'Function GetUserSecurity(securityObjectIdNo As Int32, securityGroupIdNo As Int32) As ArrayList
+    'Function GetUserSecurity(securityObjectIdNo As Int16, securityGroupIdNo As Int16) As ArrayList
 
     Function HasRecordChanged(idNo As Int32, tableName As String, timeStampedValue As Object,
                               Optional ByVal timeStampField As String = "DateTimeStamp") As Boolean
@@ -108,7 +108,11 @@ Public Interface IModel
     Sub Logout()
 
     Function GetSqlValue(Of TType)(sqlStatement As String, tableName As String, condition As String) As TType
+
     Function GetControlSecurityIdNo(searchValue As String) As String
-    Function GetUserSecurity(securityObjectIdNo As Int32, securityGroupIdNo As Int32) As ArrayList
-    Function GetUserSecurityForKey(securityObjectName As String, securityGroupIdNo As Int32) As ArrayList
+
+    Function GetUserSecurity(securityObjectIdNo As Int16, securityGroupIdNo As Int16) As ArrayList
+
+    Function GetUserSecurityForKey(securityObjectName As String, securityGroupIdNo As Int16) As ArrayList
+
 End Interface
